second commit

+++ src/main/java/com/schoolmanagement/controller/ClassController.java
@@ -3,12 +3,16 @@ package com.schoolmanagement.controller;
 
 import com.schoolmanagement.dto.ClassDTO;
 import com.schoolmanagement.entity.Class;
+import com.schoolmanagement.exceptions.ClassAlreadyExistsException;
 import com.schoolmanagement.service.ClassService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("api/classes")
@@ -19,7 +23,11 @@ public class ClassController {
     private ClassService classService;
 
     @PostMapping
-    public ResponseEntity<Class> addClass(@RequestBody Class classEntity) {
+    public ResponseEntity<Class> addClass(@RequestBody Class classEntity) throws ClassAlreadyExistsException {
+        Class fetchedClass = classService.findClassByNameAndSection(classEntity.getName(),classEntity.getSection());
+        if(fetchedClass != null) {
+            throw new ClassAlreadyExistsException("Class already exists");
+        }
         return ResponseEntity.ok(classService.addClass(classEntity));
     }
 
@@ -29,13 +37,25 @@ public class ClassController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteClass(@PathVariable Long id) {
-        classService.deleteClass(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<String> deleteClass(@PathVariable Long id) {
+        Class fetchedClass = classService.findClass(id);
+        if (fetchedClass != null) {
+            classService.deleteClass(id);
+        }else {
+            return ResponseEntity.ok("Class not found");
+        }
+        return ResponseEntity.ok("Deleted");
     }
 
     @GetMapping("/allclasses")
     public ResponseEntity<List<ClassDTO>> getAllClasses() {
         return ResponseEntity.ok(classService.getAllClasses());
+    }
+
+    @ExceptionHandler(ClassAlreadyExistsException.class)
+    public ResponseEntity<Map<String, String>> handleClassAlreadyExistsException(ClassAlreadyExistsException ex) {
+        Map<String, String> response = new HashMap<>();
+        response.put("message", ex.getMessage());
+        return new ResponseEntity<>(response, HttpStatus.CONFLICT); // 409 Conflict
     }
 }
